+++ Utils/Event.h
@@ -43,7 +43,7 @@ namespace Ph2_HwInterface {
         uint8_t fSensor;
         uint8_t fFirstStrip;
         uint8_t fClusterWidth;
-        double getBaricentre();
+        float getBaricentre();
     };
 
     class PCluster
@@ -53,7 +53,7 @@ namespace Ph2_HwInterface {
         uint8_t fAddress;
         uint8_t fWidth;
         uint8_t fZpos;
-        double getBaricentre();
+        float getBaricentre();
     };
 
     class SCluster
@@ -62,7 +62,7 @@ namespace Ph2_HwInterface {
         uint8_t fAddress;
         uint8_t fMip;
         uint8_t fWidth;
-        double getBaricentre();
+        float getBaricentre();
     };
